refactor: extract useMatrixPool hook with lazy initialization
- Rename matrixBackgroundUtils.ts to useMatrixPool.ts
- Extract shared pool generation into useMatrixPool hook
- Use lazy useRef pattern to avoid regenerating pool on every render
