+++ packages/kit/src/views/Onboardingv2/pages/MatrixBackground.native.tsx
@@ -2,28 +2,20 @@
 // text blocks instead of multiple SizableText per line. This avoids per-tick string
 // allocation and React reconciliation of 50+ text nodes, reducing Hermes GC pressure
 // during wallet creation which can cause OOM crashes on memory-constrained iOS devices.
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 
 import { SizableText } from '@onekeyhq/components';
 
-import {
-  generateMatrixPool,
-  type IMatrixBackgroundProps,
-} from './matrixBackgroundUtils';
+import { useMatrixPool, type IMatrixBackgroundProps } from './useMatrixPool';
 
 const TEXT_POOL_COUNT = 10;
 const TEXT_TOTAL_CHARS = 3000;
 
-const MatrixBackground = ({
-  updateInterval = 500,
-}: IMatrixBackgroundProps) => {
-  const poolRef = useRef<string[]>(
-    generateMatrixPool(TEXT_POOL_COUNT, TEXT_TOTAL_CHARS),
-  );
-  const [text, setText] = useState(poolRef.current[0]);
+const MatrixBackground = ({ updateInterval = 500 }: IMatrixBackgroundProps) => {
+  const pool = useMatrixPool(TEXT_POOL_COUNT, TEXT_TOTAL_CHARS);
+  const [text, setText] = useState(pool[0]);
 
   useEffect(() => {
-    const pool = poolRef.current;
     let index = 0;
     const tick = () => {
       index = (index + 1) % pool.length;
@@ -31,14 +23,10 @@ const MatrixBackground = ({
     };
     const interval = setInterval(tick, updateInterval);
     return () => clearInterval(interval);
-  }, [updateInterval]);
+  }, [pool, updateInterval]);
 
   return (
-    <SizableText
-      textAlign="center"
-      fontFamily="$monoRegular"
-      letterSpacing={2}
-    >
+    <SizableText textAlign="center" fontFamily="$monoRegular" letterSpacing={2}>
       {text}
     </SizableText>
   );
